requirement.txt
h5py
Keras
numpy
pydot
Pillow
nltk
tensorflow
pandas
re
pickle
